feat: convert to typescript
Adds gh actions, autorelease, etc.

BREAKING CHANGE: ESM only named exports no more CJS
