+++ .gitignore
@@ -1,3 +1,5 @@
 node_modules
 coverage
 .nyc_output
+package-lock.json
+dist
