+++ package.json
@@ -1,12 +1,39 @@
 {
-  "name": "multiaddr-to-uri",
+  "name": "@multiformats/multiaddr-to-uri",
   "version": "8.0.0",
   "description": "Convert a Multiaddr to a URI /dnsaddr/ipfs.io/http -> http://ipfs.io",
-  "main": "index.js",
+  "type": "module",
+  "types": "./dist/src/index.d.ts",
+  "files": [
+    "src",
+    "dist/src",
+    "!dist/test",
+    "!**/*.tsbuildinfo"
+  ],
+  "exports": {
+    ".": {
+      "import": "./dist/src/index.js"
+    }
+  },
+  "eslintConfig": {
+    "extends": "ipfs",
+    "parserOptions": {
+      "sourceType": "module"
+    }
+  },
   "scripts": {
-    "test": "npm run lint && npm run test:coverage",
-    "test:coverage": "nyc --reporter=lcov --reporter=text ava",
-    "lint": "standard"
+    "lint": "aegir lint",
+    "dep-check": "aegir dep-check dist/src/**/*.js dist/test/**/*.js",
+    "build": "tsc",
+    "pretest": "npm run build",
+    "test": "aegir test -f ./dist/test/**/*.js",
+    "test:chrome": "npm run test -- -t browser",
+    "test:chrome-webworker": "npm run test -- -t webworker",
+    "test:firefox": "npm run test -- -t browser -- --browser firefox",
+    "test:firefox-webworker": "npm run test -- -t webworker -- --browser firefox",
+    "test:node": "npm run test -- -t node --cov",
+    "test:electron-main": "npm run test -- -t electron-main",
+    "release": "semantic-release"
   },
   "keywords": [
     "multiaddr",
@@ -14,14 +41,12 @@
     "URL"
   ],
   "author": "Alan Shaw",
-  "license": "MIT",
+  "license": "Apache-2.0 or MIT",
   "devDependencies": {
-    "ava": "^3.11.1",
-    "nyc": "^15.0.0",
-    "standard": "^16.0.3"
+    "aegir": "^36.1.3"
   },
   "dependencies": {
-    "multiaddr": "^10.0.0"
+    "@multiformats/multiaddr": "^10.1.1"
   },
   "repository": {
     "type": "git",
